requirements.txt
lxml_html_clean==0.2.2
selenium==4.24.0
requests_html==0.10.0
beautifulsoup4==4.12.3
langchain_community==0.2.5
langchain==0.2.5
langchain_experimental==0.3.0
python_dotenv==1.0.1
chromadb==0.5.5
rank_bm25==0.2.2
langchain_google_genai==1.0.10
tiktoken==0.7.0
streamlit==1.38.0
db-sqlite3==0.0.1